Add massaging back to graphql function
