+++ site/gatsby-site/netlify/functions/graphql.ts
@@ -4,9 +4,8 @@ import { ApolloServer } from '@apollo/server';
 import config from '../../server/config';
 import { MongoClient } from 'mongodb';
 import { startServerAndCreateLambdaHandler, handlers } from '@as-integrations/aws-lambda';
-import cookie from 'cookie';
 import { wrapHandler } from '../../sentry-instrumentation';
-import { Handler } from '@netlify/functions';
+import { netlifyEventToLambdaEvent } from '../../src/utils/serverless';
 
 const server = new ApolloServer({
     schema,
@@ -16,8 +15,14 @@ const client = new MongoClient(config.API_MONGODB_CONNECTION_STRING);
 
 const handler = startServerAndCreateLambdaHandler(
     server,
-    handlers.createAPIGatewayProxyEventV2RequestHandler(), { context: ({ event }) => context({ req: event as any, client }) }
-);
+    handlers.createAPIGatewayProxyEventV2RequestHandler(), {
+    context: ({ event }) => {
+
+        const extendedEvent = netlifyEventToLambdaEvent(event as any);
+
+        return context({ req: extendedEvent as any, client })
+    }
+});
 
 // @ts-ignore
 module.exports = { handler: wrapHandler(handler) };
